% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ts_innov.R
\name{calc.ts.innov}
\alias{calc.ts.innov}
\title{calc time series innovation}
\usage{
calc.ts.innov(a1, a2, sigma.ar2)
}
\arguments{
\item{a1}{Yule walker parameter alpha1 - how much you weight the previous lag1 data point}

\item{a2}{Yule walker parameter alpha2 - how much you weight the lag2 data point}

\item{sigma.ar2}{the standard deviation of the entire time series}
}
\value{
sigma.innov the standard deviation of the innovation of the time series
}
\description{
calc time series innovation
}
\examples{
calc.ts.innov(0.6, -0.4, 15)
}
