% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap.R
\name{comp_grow}
\alias{comp_grow}
\title{comp_grow}
\usage{
comp_grow(r, x0 = 0.1, n = 100)
}
\arguments{
\item{r}{grow rate}

\item{n}{number of generations}
}
\description{
comp_grow
}
\details{
x(n+1) = r*x(n)*(1-x(n))
}
