% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap.R
\name{logistic_map}
\alias{logistic_map}
\title{logistic_map}
\usage{
logistic_map(rmin = 0, rmax = 4, rstep = 0.1, n = 1000, x0 = runif(1,
  0, 1))
}
\arguments{
\item{rmin}{min r.}

\item{rmax}{max r.}

\item{rstep}{r step.}

\item{n}{number of iterations.}

\item{x0}{initial value.}
}
\value{
logmap object.
}
\description{
Create a logistic map with parameters.
}
