% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap.R
\name{logmap}
\alias{logmap}
\title{logmap}
\usage{
logmap(r0 = 0, r1 = 4, length.out = 1000, maxit = 100)
}
\arguments{
\item{r0}{initial value for the r parameter.}

\item{r1}{final value for the r parameter.}

\item{length.out}{increment value for the r parameter.}

\item{maxit}{maximum number of iterations returned per r value.}
}
\description{
logmap
}
