% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelbrot.R
\name{mandelbrot}
\alias{mandelbrot}
\title{mandelbrot}
\usage{
mandelbrot(x0 = -2.5, x1 = 1, y0 = -1.5, y1 = 1.5, steps = 100,
  maxit = 100, tol = 1e-06, mc.cores = parallel::detectCores())
}
\arguments{
\item{x0}{starting point in real axis.}

\item{x1}{end point in real axis.}

\item{y0}{starting point in imaginary axis.}

\item{y1}{end point in imaginary axis.}

\item{steps}{number of points in range.}

\item{maxit}{maximum number of iterations.}

\item{tol}{tolerance.}

\item{mc.cores}{number of cores for parallel processing.}
}
\description{
mandelbrot
}
