% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.logmap}
\alias{plot.logmap}
\title{plot.logmap}
\usage{
\method{plot}{logmap}(x, size = 1, color = "black", xlim = c(0, 4))
}
\arguments{
\item{x}{logmap object.}

\item{size}{point size.}

\item{color}{point color.}

\item{xlim}{limits for x-axis.}
}
\description{
Plot a logistic map object.
}
