% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_logmap}
\alias{plot_logmap}
\title{plot.logmap}
\usage{
plot_logmap(x, size = 1, color = "black", theme_void = TRUE)
}
\arguments{
\item{x}{logmap object.}

\item{size}{point size.}

\item{color}{point color.}

\item{theme_void}{logical; whether to use theme_void (default: TRUE).}
}
\description{
Plot a logistic map object.
}
