% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mandelbrot}
\alias{plot_mandelbrot}
\title{plot_mandelbrot}
\usage{
plot_mandelbrot(x, theme_void = TRUE)
}
\arguments{
\item{x}{mandelbrot set object.}

\item{theme_void}{logical; whether to use theme_void (default: TRUE).}
}
\description{
plot_mandelbrot
}
