% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trialMat_ar2_recovery.R
\name{simulate_trialMat_ar2_recovery}
\alias{simulate_trialMat_ar2_recovery}
\title{simulate_trialMat_ar2_recovery.R}
\usage{
simulate_trialMat_ar2_recovery(
  n.types,
  n.trials,
  a1,
  a2,
  ar2.mean,
  sigma.ar2,
  sigma.ar2.bound,
  n.burnin,
  n.iter
)
}
\arguments{
\item{n.types}{numerical integer the number of types of AR2 sequences you will have}

\item{n.trials}{numerical integer number of trials you want in total in your AR2 process}

\item{a1}{numerical Yule walker parameter alpha1 - how much you weight the previous lag1 data point}

\item{a2}{numerical Yule walker parameter alpha2 - how much you weight the lag2 data point}

\item{ar2.mean}{numerical integer mean of the ar2 process}

\item{sigma.ar2}{underlying standard deviation of the generative distribution}

\item{sigma.ar2.bound}{numerical acceptable tolerance for SD generated e.g within 0.01 plus or minus of gen_sd}

\item{n.burnin}{numerical the number of samples you want to throw out before sampling the AR2 process}

\item{n.iter}{numerical, number of times you want to simulate the AR2 process with a particular parameter combo before it gives up}
}
\value{
trialMat = trialMatrix of simulated experiment data as nested datasets
}
\description{
simulate_trialMat_ar2_recovery.R
}
\examples{
simulate_trialMat_ar2_recovery(n.types = 4, n.trials = 280, a1 = 0.6, a2 = -0.4, ar2.mean = 0, sigma.ar2 = 15, sigma.ar2.bound = 0.001, n.burnin = 1000, n.iter = 50000)
}
