% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_phd.R
\name{theme_phd}
\alias{theme_phd}
\title{Add theme to ggplot}
\usage{
theme_phd()
}
\description{
This function allows you to add the mae theme to your ggplot graphics. Individual theme elements can be overwritten in subsequent ggplot code.
This code was originally written by Indra Parta for the mae team https://github.com/indraparta/maecharts
}
\examples{
\dontrun{
iris |>
 ggplot() +
 geom_point(aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
 theme_mae()

MAE theme elements can be overwritten by adding additional
theme arguments after theme_mae() call

iris |>
 ggplot() +
 geom_point(aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
 theme_mae() +
 theme(legend.position = 'right')
 }
}
\keyword{theme_mae}
